Browser settings merged into Settings
