+++ src/_locales/de/messages.json
@@ -328,9 +328,6 @@
     "openInTab": {
         "message": "In Tab öffnen"
     },
-    "browserSettings": {
-        "message": "Browsereinstellungen"
-    },
     "contextMenus": {
         "message": "Kontextmenüs"
     },
