+++ src/_locales/pt_BR/messages.json
@@ -210,7 +210,7 @@
 		"message": "Dias"
 	},
 	"errorLoadingTitle": {
-		"message": "Erro ao carregar título"	
+		"message": "Erro ao carregar título"
 	},
 	"keyboardCommandFailure": {
 		"message": "Falha na tecla de atalho"
@@ -327,9 +327,6 @@
 	},
 	"openInTab": {
 		"message": "Abrir em aba"
-	},
-	"browserSettings": {
-		"message": "Configurações do Navegador"
 	},
 	"contextMenus": {
 		"message": "Menu de contexto"
