+++ src/template/appBar/adminMenuArea.html
@@ -13,13 +13,6 @@
         </div>
       </div>
     </li>
-    <li data-ui='browserSettings' class='listItem listItem--small listItem--clickable'>
-      <div class='listItem-content'>
-        <div class='listItem-title'>
-          <%= browserSettingsMessage %>
-        </div>
-      </div>
-    </li>
     <li data-ui='keyboardShortcuts' class='listItem listItem--small listItem--clickable'>
       <div class='listItem-content'>
         <div class='listItem-title'>
