--- src/template/dialog/browserSettings.html
@@ -1,26 +1,0 @@
-﻿<section class='grouping'>
-  <div class='grouping-header'>
-    <%= contextMenusMessage %>
-  </div>
-
-  <div class='grouping-content'>
-    <ul>
-      <li data-region='showTextSelectionContextMenu' class='grouping-item listItem listItem--small'></li>
-      <li data-region='showYouTubeLinkContextMenu' class='grouping-item listItem listItem--small'></li>
-      <li data-region='showYouTubePageContextMenu' class='grouping-item listItem listItem--small'></li>
-    </ul>
-  </div>
-</section>
-
-<section class='grouping'>
-  <div class='grouping-header'>
-    <%= websiteEnhancementsMessage %>
-  </div>
-
-  <div class='grouping-content'>
-    <ul>
-      <li data-region='enhanceYouTube' class='grouping-item listItem listItem--small'></li>
-      <li data-region='enhanceBeatport' class='grouping-item listItem listItem--small'></li>
-    </ul>
-  </div>
-</section>
