+++ src/template/dialog/settings.html
@@ -36,3 +36,30 @@
     </ul>
   </div>
 </section>
+
+<section class='grouping'>
+  <div class='grouping-header'>
+    <%= contextMenusMessage %>
+  </div>
+
+  <div class='grouping-content'>
+    <ul>
+      <li data-region='showTextSelectionContextMenu' class='grouping-item listItem listItem--small'></li>
+      <li data-region='showYouTubeLinkContextMenu' class='grouping-item listItem listItem--small'></li>
+      <li data-region='showYouTubePageContextMenu' class='grouping-item listItem listItem--small'></li>
+    </ul>
+  </div>
+</section>
+
+<section class='grouping'>
+  <div class='grouping-header'>
+    <%= websiteEnhancementsMessage %>
+  </div>
+
+  <div class='grouping-content'>
+    <ul>
+      <li data-region='enhanceYouTube' class='grouping-item listItem listItem--small listItem--clickable'></li>
+      <li data-region='enhanceBeatport' class='grouping-item listItem listItem--small listItem--clickable'></li>
+    </ul>
+  </div>
+</section>
